added authorise policy gates

+++ app/Http/Controllers/CartController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\CartRequest;
 use App\Http\Resources\CartResource;
 use App\Models\Cart;
 use Illuminate\Http\Request;
@@ -15,19 +16,11 @@ class CartController extends Controller
      */
     public function index()
     {
-        $carts = Cart::all();
+        $carts = auth()->user()->carts;
+
         return CartResource::collection($carts);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
 
     /**
      * Store a newly created resource in storage.
@@ -35,9 +28,9 @@ class CartController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function store(Request $request)
+    public function store(CartRequest $request)
     {
-        $cart = Cart::create($request->all());
+        $cart = auth()->user()->carts()->create($request->validated());
 
         return response()
             ->json([
@@ -51,23 +44,18 @@ class CartController extends Controller
      * Display the specified resource.
      *
      * @param  \App\Models\Cart  $cart
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function show(Cart $cart)
     {
-        //
+        $this->authorize('view', $cart);
+
+        return response()->json([
+            'data' => new CartResource($cart)
+        ], 200 );
+
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Cart  $cart
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Cart $cart)
-    {
-        //
-    }
 
     /**
      * Update the specified resource in storage.
@@ -76,12 +64,12 @@ class CartController extends Controller
      * @param  \App\Models\Cart  $cart
      * @return \Illuminate\Http\JsonResponse
      */
-    public function update(Request $request, Cart $cart)
+    public function update(CartRequest $request, Cart $cart)
     {
 
-//        $this->authorize('update' , $cart);
+        $this->authorize('update' , $cart);
 
-        $cart->update($request->all());
+        $cart->update($request->validated());
 
         return response()->json([
             'message' => 'Cart Updated',
@@ -95,8 +83,11 @@ class CartController extends Controller
      * @param  \App\Models\Cart  $cart
      * @return \Illuminate\Http\JsonResponse
      */
+
     public function destroy(Cart $cart)
     {
+        $this->authorize('delete' , $cart);
+
         $cart->delete();
 
         return response()->json([
